Initial implementation of CLI.
Commands include:

- ad-info        Show information about an advertisement from a specified publisher
- find           Find value by CID or multihash in indexer
- providers      Commands to get provider information
- spaddr         Get storage provider p2p ID and address from lotus gateway
- verify-ingest  Verifies an indexer's ingestion of multihashes read from a publisher, a CAR file or a CARv2 Index

The `ad-info` is a rename of the index-provider's `list-ads` command.

The `verify-ingest` command is an enhanced version of index-provider's `verify-ingest` command. In includes the feature requested by https://github.com/ipni/index-provider/issues/281 as well as fixing some error handling issues.

The other commands are directly from storetheindex.