Get ad entries data as needed (#13)
* Get ad entries data as needed

Advertisement entries data was being fetched along with each ad in the chain. If fetching a large number of ads this can inefficient if not all the entries are needed. Instead, only fetch the entries data for each as at the time the entries data is needed.

This also removes unnecessary code and data from the publisher client.

* Add --distance to provider command to compute distance to head advertisement